+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/edit_userinfo/UserInfoContract.java
@@ -62,7 +62,7 @@ public interface UserInfoContract {
          * @param user_ids 用户 id 数组
          * @return
          */
-        Observable<BaseJson<List<UserInfoBean>>> getUserInfo(List<Integer> user_ids);
+        Observable<BaseJson<List<UserInfoBean>>> getUserInfo(List<Long> user_ids);
 
     }
